Move split_string to utils
